+++ src/main/java/kitchenpos/dto/request/OrderTableRequest.java
@@ -1,6 +1,5 @@
 package kitchenpos.dto.request;
 
-import javax.validation.constraints.Min;
 import javax.validation.constraints.PositiveOrZero;
 
 import kitchenpos.domain.OrderTable;
