+++ src/main/java/kitchenpos/ui/OrderRestController.java
@@ -3,8 +3,6 @@ package kitchenpos.ui;
 import java.net.URI;
 import java.util.List;
 
-import javax.validation.Valid;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
